requirements.txt
flask==3.0.0
numpy
Pillow
onnxruntime
requests
